Fix 2.18 - Extraer componente de Jobs #227
