+++ frontend/src/components/service/ServiceButton.jsx
@@ -1,4 +1,4 @@
-export default function ServiceDetailsButton({ type, text, icon, onClick }) {
+export default function ServiceButton({ type, text, icon, onClick }) {
     return (
         <div className="flex justify-center">
             <button type={type} className="flex items-center gap-x-2 p-2 border rounded-lg hover:bg-gray-200 transition" onClick={onClick}>
